+++ static/index.html
@@ -9,12 +9,14 @@
 </head>
 <body style="margin: 0;">
   <header>
-    <div class="header-container">
-      <div class="logo-font">
-        <p>台北一日遊</p>
-      </div>
-      <div class="menu">
-        <button>預定行程   登入/註冊</button>
+    <div class="navbar">
+      <div class="header-container">
+        <div class="logo-font">
+          <p>台北一日遊</p>
+        </div>
+        <div class="menu">
+          <button>預定行程   登入/註冊</button>
+        </div>
       </div>
     </div>
   </header>
@@ -43,70 +45,6 @@
       <button class="scroll-btn right"><img src="/static/img/right_arrow.svg" alt="Right Arrow"></button>
     </div>
     <div class="attractions">
-      <div class="attraction-card">
-        <div class="attraction-image">
-          <img src="https://fakeimg.pl/300/">
-          <div class="attraction-name">
-            <p>名稱</p>
-          </div>
-        </div>
-        <div class="attraction-info">
-          <div class="attraction-mrt">
-            <p>捷運站名</p>
-          </div>
-          <div class="attraction-category">
-            <p>景點類別</p>
-          </div>
-        </div>
-      </div>
-      <div class="attraction-card">
-        <div class="attraction-image">
-          <img src="https://fakeimg.pl/300/">
-          <div class="attraction-name">
-            <p>名稱</p>
-          </div>
-        </div>
-        <div class="attraction-info">
-          <div class="attraction-mrt">
-            <p>捷運站名</p>
-          </div>
-          <div class="attraction-category">
-            <p>景點類別</p>
-          </div>
-        </div>
-      </div>
-      <div class="attraction-card">
-        <div class="attraction-image">
-          <img src="https://fakeimg.pl/300/">
-          <div class="attraction-name">
-            <p>名稱</p>
-          </div>
-        </div>
-        <div class="attraction-info">
-          <div class="attraction-mrt">
-            <p>捷運站名</p>
-          </div>
-          <div class="attraction-category">
-            <p>景點類別</p>
-          </div>
-        </div>
-      </div>
-      <div class="attraction-card">
-        <div class="attraction-image">
-          <img src="https://fakeimg.pl/300/">
-          <div class="attraction-name">
-            <p>名稱</p>
-          </div>
-        </div>
-        <div class="attraction-info">
-          <div class="attraction-mrt">
-            <p>捷運站名</p>
-          </div>
-          <div class="attraction-category">
-            <p>景點類別</p>
-          </div>
-        </div>
-      </div>
     </div>
   </section>
   <footer>
@@ -114,7 +52,7 @@
       <p>COPYRIGHT © 2021 台北一日遊</p>
     </div>
   </footer>
-  <script src="/static/js/index.js"></script>
-  <script src="/static/js/search.js"></script>
+  <script type="module" src="/static/js/index.js"></script>
+  <script type="module" src="/static/js/search.js"></script>
 </body>
 </html>
